requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dill==0.3.9
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
limits==4.0.1
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.3
PyMySQL==1.1.1
python-dotenv==1.0.1
ruff==0.9.2
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
